perf(chat): smooth reasoning token streaming
Reasoning/thinking parts rendered each THINKING_DELTA immediately, so
the text appeared in uneven network-sized bursts (a pulsing reveal),
unlike assistant text which already streamed through SmoothResponse.

Extract the frame-aligned reveal from SmoothResponse into a shared
useSmoothText hook and use it for both assistant text and reasoning so
thinking streams in at a steady cadence. The reasoning auto-scroll now
also follows the revealed text, keeping the tail pinned during catch-up.
